+++ SimpleGame/SimpleGame/SceneMgr.h
@@ -24,7 +24,7 @@ private :
 	void colison_test();
 	void Update_Scene();
 	int building_image[2];
-	long Bt_time, Prv_time;
+	float Bt_time, Prv_time;
 	int particle;
 	int mapimage,animation;
 	int frame, an_state;
